done customer

+++ src/customer/customer.controller.ts
@@ -1,51 +1,77 @@
 import { Request, Response } from "express";
+import { HttpResponse } from "../shared/response/http.response";
 import { CustomerService } from "./customer.service";
 
 export class CustomerController {
+
     constructor(
-        private readonly customerService: CustomerService = new CustomerService()
+        private readonly customerService: CustomerService = new CustomerService(),
+        private readonly httpResponse: HttpResponse = new HttpResponse()
     ) { }
+
     async getCustomers(req: Request, res: Response) {
         try {
             const data = await this.customerService.findAllCustomers();
-            res.status(200).json(data);
+            return this.httpResponse.Ok(res, data)
         } catch (e) {
-            console.error(e);
+            return this.httpResponse.Error(res, e)
         }
     }
+
     async getCustomerById(req: Request, res: Response) {
         const { id } = req.params;
         try {
             const data = await this.customerService.findCustomerById(id);
-            res.status(200).json(data);
+            if (data)
+                return this.httpResponse.Ok(res, data)
+            return this.httpResponse.NotFound(res, "Customer not found")
         } catch (e) {
-            console.error(e);
+            return this.httpResponse.Error(res, e)
         }
     }
+
     async createCustomer(req: Request, res: Response) {
         try {
             const data = await this.customerService.createCustomer(req.body);
-            res.status(200).json(data);
+            return this.httpResponse.Ok(res, data)
         } catch (e) {
-            console.error(e);
+            return this.httpResponse.Error(res, e)
         }
     }
+
     async updateCustomer(req: Request, res: Response) {
         const { id } = req.params;
         try {
-            const data = await this.customerService.updateCustomer(id, req.body);
-            res.status(200).json(data);
+            const isExist = await this.customerService.findCustomerById(id)
+            if (!isExist) {
+                return this.httpResponse.NotFound(res, "Customer not found");
+            }
+            const isUpdated = await this.customerService.updateCustomer(id, req.body);
+            if (!isUpdated.affected) {
+                return this.httpResponse.Error(res, "Something went wrong");
+            }
+            const data = await this.customerService.findCustomerById(id)
+            return this.httpResponse.Ok(res, data)
         } catch (e) {
-            console.error(e);
+            return this.httpResponse.Error(res, e)
         }
     }
+
     async deleteCustomer(req: Request, res: Response) {
         const { id } = req.params;
         try {
-            const data = await this.customerService.deleteCustomer(id);
-            res.status(200).json(data);
+            const isExist = await this.customerService.findCustomerById(id)
+            if (!isExist) {
+                return this.httpResponse.NotFound(res, "Customer not found");
+            }
+            const isDeleted = await this.customerService.deleteCustomer(id);
+            if (!isDeleted.affected) {
+                return this.httpResponse.Error(res, "Something went wrong");
+            }
+            return this.httpResponse.Ok(res, isExist)
         } catch (e) {
-            console.error(e);
+            return this.httpResponse.Error(res, e)
         }
     }
+
 }
